Fix image header initialization constant typo
There as a multi-character character constant in version of the constant used for marking the image header. The version of GCC used a the time appears to have ignored the extra character silently. More recent versions of GCC end up effectively taking the value of the last character in the constant. Fix the constants so the header is created as intended.